Pyvo for returned dataset (with get=record)
When getting the record of the query, we get a dict of datasets indexed by their ids. (id, dataset) tuple is of type (str, Dataset) where Dataset is defined in mocserver.dataset. On the dataset object, it is possible to perform specific searches such as a SimpleConeSearch, a TapSearch, SSA, SIA, ... . We then get a votable containing sources matching the search.